pyproject.toml
[project]
name = "towr-casadi"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "casadi==3.6.4",
    "matplotlib==3.8.0",
    "numpy==1.26.2",
    "path-dict==4.0.0",
    "pyyaml>=6.0.3",
    "rich==13.7.0",
    "scienceplots==2.1.0",
    "scipy==1.11.4",
    "vedo==2023.4.6",
    "vtk==9.3.0",
]




# ---- Build system ----

# Dynamic versioning from:
# https://github.com/ninoseki/uv-dynamic-versioning/

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
# The source location for the package.
#packages = ["src/{{ package_module }}"]
packages = [
    'centroidal_walk',
    'centroidal_walk.collocation',
    'centroidal_walk.collocation.foot_trajectory',
    'centroidal_walk.collocation.opti',
    'centroidal_walk.collocation.opti.casadi_util',
    'centroidal_walk.collocation.spline',
    'centroidal_walk.visualization',
]